fix: extract error code to error.ts
